BackEnd/requirements.txt
Django>=5.1,<7.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.4.0
django-filter>=24.2
drf-spectacular>=0.27.2
psycopg2-binary>=2.9.9
reportlab>=4.2.2
python-decouple>=3.8
